+++ CHANGELOG.md
@@ -1,8 +1,17 @@
+## 15.0.0-beta.1 (2023-11-15)
+### Features
+* Bring OntimizeKeycloakMultitenantModule module from release 8.2.0 ([0cab31f](https://github.com/OntimizeWeb/ontimize-web-ngx-keycloak/commit/0cab31f))
+
 ## 15.0.0-beta.0 (2023-09-29)
 ### Features
 * Migration to Angular 15.
 * Migration to Angular Material 15.
-## 8.1.0 (2022-02-27)
+
+## 8.2.0 (2023-07-06)
+### Features
+* **Multitenant support**: added support for authenticating with Keycloak multitenants for OntimizeWeb applications. ([be855c9](https://github.com/OntimizeWeb/ontimize-web-ngx-keycloak/commit/be855c93116812388346fdf451081b2cf65bc717))
+
+## 8.1.0 (2023-02-27)
 ### Features
 * **OKeycloakAuthService**: `keycloakService` attribute is now public ([71f0e0c](https://github.com/OntimizeWeb/ontimize-web-ngx-keycloak/commit/71f0e0c))